Move utility functions to scripteditor/utils.py
